clean up bivar
